zcatch: reset spree on round start (Closed #137)
The current spree is intentionally not reset in ddnet-insta
to allow collecting best spree records without being interruped by round
ends.
Otherwise players are limited by how long the round is to get a higher
spree record. This would encourage prolonging the game if you want to
get the highest spree. Which is not good in team based gamemodes.

This is probelmatic in zCatch because we use the current spree to
determine the skin color. So in zCatch you are now limited to sprees
tracking per round. Which should be fine because prolonging a zCatch
game is not causing any harm for others. If a player wants to farm the
best spree in zCatch they are now forced to make use of the release
feature which is totally fine.