fng/stats: track fng freezes in base stats shots_hit
Closed  #150

Is hammer accuracy actually interesting? Weapon shots and hits are all tracked together in the base stats. I don't think a bored air hammer should affect the overall hit rate stats. It is not really a weapon where aim is interesting. For all current gametypes I would say it is not interesting.

Also it is called shots_fired and shots_hit. One does not shoot a hammer.

Whoever wants to track hammer hits for their gamemode such as zCatch hammer or bomb can do so in their gamemode. But the base stats will exclude hammer hits. And I am not planning to track it in fng either.