Remove ``sv_fastcap`` and always track flag times
Flag times are now cleanly implemented in the background.
Without ever printing weird finish messages in the chat or messing with
the game timer.

There is no need to turn it off because it can not possibly bother a
player. The only way to know about fast caps is if you query the ranks.

All the things such as race timers and ghosts are out of scope.
This is a pvp modification after all.