zcatch: never get caught by dead players
This helps a bit with #160
It does not full solve #160 because the player can respawn before its projectile hit.

Still this seems like a sensible thing to do either way to avoid bugs.
Because dead players can not die. So if you ever end up in that state
you are stuck.