Move bang commands back to gamecontext
I moved the bang and chat commands of ddnet-insta from gamecontext
to gamecontroller and back multiple times now.

Now that ddnet in issue 7777 officially decided to not support custom
forks. I do not see a way to make chat commands in the gamecontroller
work.

The current chat command state is clean to maintain and has no git
conflict surface to ddnet. But it is also not using ddnets console
system. So it does not properly check for duplicated commands.
It also does not send auto completions to the clients.
It also does not use ddnet's argument parser.

So the plan now is moving back to ddnets system in gamecontext.
Maybe one day there could be some macro magic that allows gametypes to
add new commands into gamecontext without having to leave the gametype
directory.

Another advantage of using gamecontexts Register() methods is that the
get a help text that can be used to auto generate documentation
See #69