zcatch: change skin color even if stat track is off
Closed #148

There are now two spree counters. One for stat track on and one for off. And zCatch uses the one thats currently being incremented for the colors. So if players join or leave your skin color switches from one spree counter to another.

This is a bit of a weird feature and might be interpreted as bug by users.

But this way you can update your skin color even when playing with your dummy. But you can not farm a skin color with a dummy and then keep it when more players connect.