Increase json round stats buffer even more
Makes sure that 10v10 to http as json always works.
The discord buffer is lower because its two and the result is
usually monitored by humans and not in json.

Anything above 10v10 does not seem worth the ram waste.
I don't think anyone is seriously interested in round stats of games
that big.