vanilla: turn off ddrace tile prediction to fix sg for specs
If the server is in tournament mode or the in game slots are full.
And a new player with ddnet client connects to spectators.
The shotgun and gun tuning is wrong. It is using the hardcodet ddrace
tuning instead of the vanilla tuning until the client joins into the
game.

Removing the ddrace tile prediction flag fixes that. But it also
breaks stoppa prediction and other ddrace tiles.

That is sad but having the core weapons predicted correctly also for
spectators is more important.

A proper fix depends on the ddnet client to be more flexible see
https://github.com/ddnet/ddnet/issues/8923

Closed #181