fix: disallow showing hook collision of others
The setting ``cl_show_hook_coll_other 2`` displays the hook line
of all players no matter if they press it or not at all times.

This could be used to reveal their location. This is now no longer
possible. You can still see the hookline of others if they manually
turn it on.

Closed #195