zcatch: require 5 players to play a real game
Simplifies the code. There is no more running vs running competitive.
There is now only running or not.
The amount of players is not configurable!

Below 5 players is always a release game. Above 5 players can start a
real game that once started will count as a real win no matter with
how many kills or players the game is ended.

Users can still manually vote for release games if they wish to
play gdm with 10 players.

Reverts #124
Releated to #216