zcatch: forced spec depends only on kills that count
We can still join the game if someone is on a killing spree
with kills that do not count. For example released kills.

The only thing that matters for forced spec are kills that count.
This commit fixes a bug where players end up in spec on join
during a release game.